Big rewrite of palette swapping
DMA now properly works without dots, required large rewrite and manual assembly. The timing is critical, so having values preloaded in registers is absolutely imperative (even though gcc really doesn't want to do that)

VBlank undoing of palette swap replaced with a second HInt swap, with callback setup to handle that Allows positioning the text frame at any height in the screen.

SetPalette_Fast is a big mess right now but will be better to clean up next commit, for now want a commit a that leaves all the mess in for reference later if it's ever needed