Various character render fixes
If character is not 320 pixels wide, they're now correctly centred (can specify tile offset per animation frame).
When animation is changed, the first frame now renders straight away as the animator expects it.
Only clear the dialogue mode if we're going from no character shown to a character, or vice versa.
Add SetMap_SubFull that uploads a tilemap smaller than the screen row-by-row, only in the region it will affect. Use it to apply characters