Set up installation for sovereign
Web target of adventure-md and sovereign added. It's the sort of thing it'd be better to have as a configuration in CMake, but I pretty much always just want to click one button to install everything - that's the whole point of doing this work! So it's two targets in every config, and a new config specifically for installations.

Installing, creates the folder "distribute" and puts the MD rom, win/mac/linux zips with blastem and the MD rom, and a web zip with the web rom.

Emu installation requires blastem builds to be placed in folders in distribute_src. Could in principle set up aa download step but that's not worth it for me rn.